+++ common/src/main/java/net/spell_power/api/DamageCurve.java
@@ -1,3 +1,3 @@
 package net.spell_power.api;
 
-public enum DamageCurve {LINEAR, QUADRATIC, HYPERBOLIC}
+public enum DamageCurve { LINEAR, QUADRATIC, HYPERBOLIC }