+++ common/src/main/java/net/spell_power/api/ModifierDefinitions.java
@@ -1,0 +1,8 @@
+package net.spell_power.api;
+
+import net.minecraft.util.Identifier;
+import net.spell_power.SpellPowerMod;
+
+public class ModifierDefinitions {
+    public static Identifier INNATE_BONUS = Identifier.of(SpellPowerMod.ID, "innate_bonus");
+}