+++ common/src/main/java/net/spell_power/config/AttributesConfig.java
@@ -1,8 +1,6 @@
 package net.spell_power.config;
 
 import net.spell_power.api.DamageCurve;
-import net.spell_power.api.SpellPowerMechanics;
-import net.spell_power.internals.SpellStatusEffect;
 
 import java.util.Map;
 
@@ -13,8 +11,8 @@ public class AttributesConfig {
     public AttributeScope attributes_container_injection_scope = AttributeScope.LIVING_ENTITY;
     public double base_spell_critical_chance_percentage = 5;
     public double base_spell_critical_damage_percentage = 50;
-    public SpellStatusEffect.Config spell_power_effect = new SpellStatusEffect.Config(0.1F);
-    public Map<String, SpellStatusEffect.Config> secondary_effects;
+    public EffectConfig spell_power_effect = new EffectConfig(0.1F);
+    public Map<String, EffectConfig> secondary_effects;
 
 
     public DamageCurve resistance_curve = DamageCurve.HYPERBOLIC;
@@ -26,9 +24,13 @@ public class AttributesConfig {
     public static AttributesConfig defaults() {
         var config = new AttributesConfig();
         config.secondary_effects = Map.of(
-                SpellPowerMechanics.CRITICAL_CHANCE.name, new SpellStatusEffect.Config(0.05F),
-                SpellPowerMechanics.CRITICAL_DAMAGE.name, new SpellStatusEffect.Config(0.1F),
-                SpellPowerMechanics.HASTE.name, new SpellStatusEffect.Config(0.05F)
+                "critical_chance", new EffectConfig(0.05F),
+                "critical_damage", new EffectConfig(0.1F),
+                "haste", new EffectConfig(0.05F)
+// Disabled due to static init circularity
+//                SpellPowerMechanics.CRITICAL_CHANCE.name, new SpellStatusEffect.Config(0.05F),
+//                SpellPowerMechanics.CRITICAL_DAMAGE.name, new SpellStatusEffect.Config(0.1F),
+//                SpellPowerMechanics.HASTE.name, new SpellStatusEffect.Config(0.05F)
         );
         return config;
     }
@@ -52,5 +54,15 @@ public class AttributesConfig {
         }
 
         return true;
+    }
+
+    public static class EffectConfig {
+        public float bonus_per_stack = 0.1F;
+
+        public EffectConfig() { }
+
+        public EffectConfig(float bonus_per_stack) {
+            this.bonus_per_stack = bonus_per_stack;
+        }
     }
 }